split get step values into seperate file
